feat: moved older style of chest to own files to avoid merge issues
